Proposal of recording cams streams
Please test it!
First, from Add-on Settings->General, set the path where the stream will be saved (same as tha path for downloads) and the location of the ffmpeg binary (you have to download it first!)
Recordings are made as you watch and stops when you stop watching.
There are 2 options: from contextual menu "Download video". The second, in Add-on settings you set "Download stream" to true, so on any model you enter the recording will start.
The file name is the name of the model+datetime. The container is mkv